requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.4
click==8.0.1
colorful==0.5.4
Flask==2.0.1
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
meraki==1.10.0
multidict==5.1.0
prettyprinter==0.18.0
Pygments==2.9.0
python-dotenv==0.19.0
requests==2.26.0
typing-extensions==3.10.0.0
urllib3==1.26.6
Werkzeug==2.0.1
yarl==1.6.3
gunicorn==20.1.0
Flask-Session
